backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
praw==7.7.1
joblib==1.3.2
pandas==2.1.4
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.4
gunicorn==21.2.0
requests==2.31.0
Werkzeug==2.3.7
urllib3==1.26.15
